nb v0.1.0 Nb.Events

Wraps create, update, and get of Nationbuilder’s Events API

Functions

create(event, site)

create(binary, binary) :: map

POSTs the event to the create endpoint

stream_all(site)

stream_all(binary) :: list

Streams all events for a given site Filters out unlisted events, and events without addresses

stream_for(calendar_id, site)

stream_for(binary, binary) :: list

Streams all events for a given calendar and site Filters out unlisted events, and events without addresses

update(id, event, site)

update(binary, map, binary) :: map

POSTs the event to the update endpoint for the event id