nb v0.1.0 Nb.People

Functions

add_tags(id, tags)

add_tags(binary, [binary]) :: map

Adds all tags (tags can be a string or an array of strings) to the person with id id

delete_tags(id, tags)

delete_tags(binary, [binary]) :: map

Deletes all tags (tags can be a string or an array of strings) to the person with id id

match(body)

match(map) :: map

Converts body to a query string and hits /people/match Returns either the person resource, or {:error, error_message}

push(person)

push(map) :: map

PUTs person to /people/push

show(id)

show(binary) :: map

Gets the full person resource – GET /person/:id

update(id, updates)

update(binary, map) :: map

PUTs updates to /people/